Recover fn_800D8414 to 100% (drop redundant (s16) cast on already-s16 gPlayerMoveTargetYaw; extsh lands directly in diff's r4 instead of a fresh r3 temp)

+++ src/main/dll/dll_000F_unk.c
@@ -428,7 +428,7 @@ void fn_800D8414(int* obj, int* ctx)
     ((BaddieState*)ctx)->inputMagnitude = ((BaddieState*)ctx)->inputMagnitude / lbl_803E0578;
     gPlayerMoveTargetYaw = getAngle(((BaddieState*)ctx)->moveInputX, -((BaddieState*)ctx)->moveInputZ);
     gPlayerMoveTargetYaw -= ((BaddieState*)ctx)->cameraYaw;
-    diff = (s16)gPlayerMoveTargetYaw;
+    diff = gPlayerMoveTargetYaw;
     diff -= (u16)((GameObject*)obj)->anim.rotX;
     if (diff > 0x8000)
     {
